初始化

+++ dms/dms-smf-spring-boot-autoconfigure/src/main/java/com/shudun/dms/rpc/MessageFuture.java
@@ -19,7 +19,7 @@ public class MessageFuture {
 
     private Message message;
 
-    private int errCode;
+    private long errCode;
 
     private long timeout;
